moved the port waiter to configuration folder.
